Check if the inputDirectory exists before run the plugin Started support for subfolders

+++ src/main/java/com/ruleoftech/markdown/page/generator/plugin/MdPageGeneratorMojo.java
@@ -39,26 +39,41 @@ public class MdPageGeneratorMojo extends AbstractMojo {
 	@Parameter(property = "generate.footerHtmlFile")
 	private String footerHtmlFile;
 
+	@Parameter(property = "generate.recursiveInput")
+	private boolean recursiveInput;
+
 	// Possible options
-	// SMARTS: Beautifies apostrophes, ellipses ("..." and ". . .") and dashes ("--" and "---")
-	// QUOTES: Beautifies single quotes, double quotes and double angle quotes (« and »)
-	// SMARTYPANTS: Convenience extension enabling both, SMARTS and QUOTES, at once.
+	// SMARTS: Beautifies apostrophes, ellipses ("..." and ". . .") and dashes
+	// ("--" and "---")
+	// QUOTES: Beautifies single quotes, double quotes and double angle quotes
+	// (« and »)
+	// SMARTYPANTS: Convenience extension enabling both, SMARTS and QUOTES, at
+	// once.
 	// ABBREVIATIONS: Abbreviations in the way of PHP Markdown Extra.
-	// HARDWRAPS: Alternative handling of newlines, see Github-flavoured-Markdown
-	// AUTOLINKS: Plain (undelimited) autolinks the way Github-flavoured-Markdown implements them.
-	// TABLES: Tables similar to MultiMarkdown (which is in turn like the PHP Markdown Extra tables, but with colspan support).
+	// HARDWRAPS: Alternative handling of newlines, see
+	// Github-flavoured-Markdown
+	// AUTOLINKS: Plain (undelimited) autolinks the way
+	// Github-flavoured-Markdown implements them.
+	// TABLES: Tables similar to MultiMarkdown (which is in turn like the PHP
+	// Markdown Extra tables, but with colspan support).
 	// DEFINITION LISTS: Definition lists in the way of PHP Markdown Extra.
-	// FENCED CODE BLOCKS: Fenced Code Blocks in the way of PHP Markdown Extra or Github-flavoured-Markdown.
+	// FENCED CODE BLOCKS: Fenced Code Blocks in the way of PHP Markdown Extra
+	// or Github-flavoured-Markdown.
 	// HTML BLOCK SUPPRESSION: Suppresses the output of HTML blocks.
 	// INLINE HTML SUPPRESSION: Suppresses the output of inline HTML elements.
-	// WIKILINKS Support [[Wiki-style links]] with a customizable URL rendering logic.
+	// WIKILINKS Support [[Wiki-style links]] with a customizable URL rendering
+	// logic.
 	@Parameter(property = "generate.pegdownExtensions", defaultValue = "TABLES")
 	private String pegdownExtensions;
 
 	private enum EPegdownExtensions {
-		NONE(0x00), SMARTS(0x01), QUOTES(0x02), SMARTYPANTS(EPegdownExtensions.SMARTS.getValue() + EPegdownExtensions.QUOTES.getValue()), ABBREVIATIONS(
-				0x04), HARDWRAPS(0x08), AUTOLINKS(0x10), TABLES(0x20), DEFINITIONS(0x40), FENCED_CODE_BLOCKS(0x80), WIKILINKS(0x100), ALL(
-				0x0000FFFF), SUPPRESS_HTML_BLOCKS(0x00010000), SUPPRESS_INLINE_HTML(0x00020000), SUPPRESS_ALL_HTML(0x00030000);
+		NONE(0x00), SMARTS(0x01), QUOTES(0x02), SMARTYPANTS(
+				EPegdownExtensions.SMARTS.getValue()
+						+ EPegdownExtensions.QUOTES.getValue()), ABBREVIATIONS(
+				0x04), HARDWRAPS(0x08), AUTOLINKS(0x10), TABLES(0x20), DEFINITIONS(
+				0x40), FENCED_CODE_BLOCKS(0x80), WIKILINKS(0x100), ALL(
+				0x0000FFFF), SUPPRESS_HTML_BLOCKS(0x00010000), SUPPRESS_INLINE_HTML(
+				0x00020000), SUPPRESS_ALL_HTML(0x00030000);
 
 		private final int value;
 
@@ -91,18 +106,21 @@ public class MdPageGeneratorMojo extends AbstractMojo {
 		if (StringUtils.isNotEmpty(copyDirectories)) {
 			getLog().info("Copy files from directories");
 			for (String dir : copyDirectories.split(",")) {
-				copyFiles(inputDirectory + File.separator + dir, outputDirectory + File.separator + dir);
+				copyFiles(inputDirectory + File.separator + dir,
+						outputDirectory + File.separator + dir);
 			}
 		}
 
 		getLog().info("Read Markdown files from input directory");
-		readFiles();
+		boolean hasFiles = readFiles();
 
-		getLog().info("Process Pegdown extension options");
-		int options = getPegdownExtensions(pegdownExtensions);
+		if (hasFiles) {
+			getLog().info("Process Pegdown extension options");
+			int options = getPegdownExtensions(pegdownExtensions);
 
-		getLog().info("Parse Markdown to HTML");
-		processMarkdown(markdownDTOs, options);
+			getLog().info("Parse Markdown to HTML");
+			processMarkdown(markdownDTOs, options);
+		}
 	}
 
 	private int getPegdownExtensions(String extensions) {
@@ -116,7 +134,9 @@ public class MdPageGeneratorMojo extends AbstractMojo {
 			} catch (NoSuchFieldException e) {
 				throw new IllegalArgumentException("No such extension: " + ext);
 			} catch (IllegalAccessException e) {
-				throw new IllegalArgumentException("Cannot read int value for extension " + ext + ": " + e, e);
+				throw new IllegalArgumentException(
+						"Cannot read int value for extension " + ext + ": " + e,
+						e);
 			}
 		}
 
@@ -132,21 +152,31 @@ public class MdPageGeneratorMojo extends AbstractMojo {
 	 * @throws MojoExecutionException
 	 *             Unable to load file
 	 */
-	private void readFiles() throws MojoExecutionException {
+	private boolean readFiles() throws MojoExecutionException {
 		getLog().debug("Read files from: " + inputDirectory);
 
 		try {
+			File inputFile = new File(inputDirectory);
+
+			if (!inputFile.exists()) {
+				getLog().info(
+						"There is no input folder for the project. Skipping.");
+				return false;
+			}
 			// Reading just the markdown dir and not sub dirs
-			List<File> markdownFiles = getFilesAsArray(FileUtils.iterateFiles(new File(inputDirectory), new String[] { "md" }, false));
+			List<File> markdownFiles = getFilesAsArray(FileUtils.iterateFiles(
+					inputFile, new String[] { "md" }, recursiveInput));
 
 			for (File file : markdownFiles) {
 				getLog().debug("File getName() " + file.getName());
-				getLog().debug("File getAbsolutePath() " + file.getAbsolutePath());
+				getLog().debug(
+						"File getAbsolutePath() " + file.getAbsolutePath());
 				getLog().debug("File getPath() " + file.getPath());
 
 				MarkdownDTO dto = new MarkdownDTO();
 				dto.markdownFile = file;
 
+				// TODO Fix copy and include the folder structure
 				if (!StringUtils.isNotEmpty(defaultTitle)) {
 					List<String> raw = FileUtils.readLines(file);
 					dto.title = getTitle(raw);
@@ -154,24 +184,28 @@ public class MdPageGeneratorMojo extends AbstractMojo {
 					dto.title = defaultTitle;
 				}
 
-				File htmlFile = new File(outputDirectory + "/" + file.getName().replaceAll(".md", ".html"));
+				File htmlFile = new File(outputDirectory + "/"
+						+ file.getName().replaceAll(".md", ".html"));
 				dto.htmlFile = htmlFile;
 
 				markdownDTOs.add(dto);
 			}
 		} catch (IOException e) {
-			throw new MojoExecutionException("Unable to load file " + e.getMessage(), e);
+			throw new MojoExecutionException("Unable to load file "
+					+ e.getMessage(), e);
 		}
+		return true;
 	}
 
 	/**
-	 * Going through list of DTOs and parsing the markdown into HTML.
-	 * Add header and footer to the big String.
+	 * Going through list of DTOs and parsing the markdown into HTML. Add header
+	 * and footer to the big String.
 	 * 
 	 * @throws MojoExecutionException
 	 *             Unable to write file
 	 */
-	private void processMarkdown(List<MarkdownDTO> markdownDTOs, int options) throws MojoExecutionException {
+	private void processMarkdown(List<MarkdownDTO> markdownDTOs, int options)
+			throws MojoExecutionException {
 		getLog().debug("processMarkdown");
 
 		for (MarkdownDTO dto : markdownDTOs) {
@@ -181,16 +215,19 @@ public class MdPageGeneratorMojo extends AbstractMojo {
 				String headerHtml = "";
 				String footerHtml = "";
 				if (StringUtils.isNotEmpty(headerHtmlFile)) {
-					headerHtml = FileUtils.readFileToString(new File(headerHtmlFile));
+					headerHtml = FileUtils.readFileToString(new File(
+							headerHtmlFile));
 					headerHtml = addTitleToHtmlFile(headerHtml, dto.title);
 				}
 				if (StringUtils.isNotEmpty(footerHtmlFile)) {
-					footerHtml = FileUtils.readFileToString(new File(footerHtmlFile));
+					footerHtml = FileUtils.readFileToString(new File(
+							footerHtmlFile));
 				}
 				String markdown = FileUtils.readFileToString(dto.markdownFile);
 				// getLog().debug(markdown);
 
-				String markdownAsHtml = new PegDownProcessor(options).markdownToHtml(markdown);
+				String markdownAsHtml = new PegDownProcessor(options)
+						.markdownToHtml(markdown);
 				StringBuilder data = new StringBuilder();
 				data.append(headerHtml);
 				data.append(markdownAsHtml);
@@ -199,7 +236,8 @@ public class MdPageGeneratorMojo extends AbstractMojo {
 				FileUtils.writeStringToFile(dto.htmlFile, data.toString());
 			} catch (IOException e) {
 				getLog().error("Error : " + e.getMessage(), e);
-				throw new MojoExecutionException("Unable to write file " + e.getMessage(), e);
+				throw new MojoExecutionException("Unable to write file "
+						+ e.getMessage(), e);
 			}
 		}
 	}
@@ -251,17 +289,29 @@ public class MdPageGeneratorMojo extends AbstractMojo {
 	 * @throws MojoExecutionException
 	 *             Unable to copy file
 	 */
-	private void copyFiles(String fromDir, String toDir) throws MojoExecutionException {
+	private void copyFiles(String fromDir, String toDir)
+			throws MojoExecutionException {
 		getLog().debug("fromDir=" + fromDir + "; toDir=" + toDir);
 		try {
 			// get files but not from the sub dirs
-			Iterator<File> files = FileUtils.iterateFiles(new File(fromDir), null, false);
-			while (files.hasNext()) {
-				File file = files.next();
-				FileUtils.copyFileToDirectory(file, new File(toDir));
+			File fromDirFile = new File(fromDir);
+			if (fromDirFile.exists()) {
+				Iterator<File> files = FileUtils.iterateFiles(fromDirFile,
+						null, false);
+				while (files.hasNext()) {
+					File file = files.next();
+					if (file.exists()) {
+						FileUtils.copyFileToDirectory(file, new File(toDir));
+					} else {
+						getLog().error(
+								"File '" + file.getAbsolutePath()
+										+ "' does not exist. Skipping copy");
+					}
+				}
 			}
 		} catch (IOException e) {
-			throw new MojoExecutionException("Unable to copy file " + e.getMessage(), e);
+			throw new MojoExecutionException("Unable to copy file "
+					+ e.getMessage(), e);
 		}
 	}
